+++ src/test/scala/scorex/waves/NodeAPISpecification.scala
@@ -3,9 +3,8 @@ package scorex.waves
 import org.scalatest.{BeforeAndAfterAll, DoNotDiscover, FunSuite, Matchers}
 import scorex.api.http.ApiKeyNotValid
 
-@Deprecated
 @DoNotDiscover
-class ScorexAPISpecification extends FunSuite with Matchers with BeforeAndAfterAll {
+class NodeAPISpecification extends FunSuite with Matchers with BeforeAndAfterAll {
 
   import TestingCommons._
 
@@ -17,21 +16,21 @@ class ScorexAPISpecification extends FunSuite with Matchers with BeforeAndAfterA
     stop()
   }
 
-  test("/scorex/status API route") {
-    val status = getRequest("/scorex/status")
+  test("/node/status API route") {
+    val status = getRequest("/node/status")
     List("generating", "syncing") should contain((status \ "blockGeneratorStatus").as[String])
     List("synced", "syncing", "idle") should contain((status \ "historySynchronizationStatus").as[String])
   }
 
-  test("/scorex/version API route") {
-    val version = getRequest("/scorex/version")
+  test("/node/version API route") {
+    val version = getRequest("/node/version")
     (version \ "version").as[String].contains("Waves") shouldBe true
     (version \ "version").as[String].contains("Release0") shouldBe true
     (version \ "version").as[String].contains("v0.2.") shouldBe true
   }
 
-  test("/scorex/stop API route protected by api key") {
-    val wrongKeyResponse = postRequest(us = "/scorex/stop", headers = Map("api_key" -> "wrong")).toString
+  test("/node/stop API route protected by api key") {
+    val wrongKeyResponse = postRequest(us = "/node/stop", headers = Map("api_key" -> "wrong")).toString
     assert(wrongKeyResponse == ApiKeyNotValid.json.toString(), s"$wrongKeyResponse == ${ApiKeyNotValid.json.toString()} is false")
   }
 }